+++ ft_printf/srcs/ft_printf/ft_printf_get_flags.c
@@ -6,7 +6,7 @@
 /*   By: bopopovi <bopopovi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/29 19:03:18 by bopopovi          #+#    #+#             */
-/*   Updated: 2018/09/06 16:46:10 by bopopovi         ###   ########.fr       */
+/*   Updated: 2018/09/06 19:15:18 by bopopovi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -89,7 +89,7 @@ static void	init_flags(t_ptf *ptf)
 ** Returns index at which function stopped
 */
 
-int			ft_printf_get_flags(t_ptf *ptf, va_list ap, int i)
+uint64_t		ft_printf_get_flags(t_ptf *ptf, va_list ap, uint64_t i)
 {
 	const char	*fmt;
 	char		*flags;
@@ -100,9 +100,9 @@ int			ft_printf_get_flags(t_ptf *ptf, va_list ap, int i)
 	while (fmt[i] && ft_printf_is_flag(fmt[i]))
 	{
 		if (fmt[i] == '.')
-			i += get_precision(ap, fmt + i + 1, &(ptf->precision)) + 1;
+			i += (uint64_t)get_precision(ap, fmt + i + 1, &ptf->precision) + 1;
 		else if (ft_strchr("123456789*", fmt[i]))
-			i += get_width(ap, fmt + i, &(ptf->width), flags);
+			i += (uint64_t)get_width(ap, fmt + i, &(ptf->width), flags);
 		else if ((ft_strchrn(flags, (int)fmt[i])) < 2)
 			ft_strncat(flags, &fmt[i], 1);
 		i++;
